refactor(functons): gether handlers
