--- packages/functions/src/functions/auth/onDeleteUser.ts
@@ -1,6 +1,0 @@
-import { setUserToNotActive } from '../../utils/firestore/users';
-import { AuthOnDeleteHandler } from '../../types/functions';
-
-export const onDeleteUserHandler: AuthOnDeleteHandler = async ({ uid }) => {
-  await setUserToNotActive(uid);
-};
