+++ packages/functions/src/handlers/getFollowers.ts
