+++ packages/functions/src/handlers/onCreateWatch.ts
@@ -2,7 +2,7 @@ import { FirestoreDateLike, WatchData, RecordData, RecordUserData } from '@yukuk
 import * as functions from 'firebase-functions';
 import * as Twitter from 'twitter';
 import * as _ from 'lodash';
-import { firestore } from '../../modules/firebase';
+import { firestore } from '../modules/firebase';
 import {
   checkNoUserMatches,
   checkInvalidToken,
@@ -10,13 +10,13 @@ import {
   getToken,
   existsRecords,
   setTwUsers,
-} from '../../utils/firestore';
-import { getTwUser } from '../../utils/firestore/twUsers/getTwUser';
-import { addRecords } from '../../utils/firestore/records/addRecords';
-import { addRecord } from '../../utils/firestore/records/addRecord';
-import { getUsersLookup } from '../../utils/twitter';
-import { mergeWatches } from '../../utils/watches';
-import { FirestoreOnCreateHandler } from '../../types/functions';
+} from '../utils/firestore';
+import { getTwUser } from '../utils/firestore/twUsers/getTwUser';
+import { addRecords } from '../utils/firestore/records/addRecords';
+import { addRecord } from '../utils/firestore/records/addRecord';
+import { getUsersLookup } from '../utils/twitter';
+import { mergeWatches } from '../utils/watches';
+import { FirestoreOnCreateHandler } from '../types/functions';
 
 const emptyRecord: RecordData<FirestoreDateLike> = {
   type: 'kuru',
