+++ packages/functions/src/handlers/onDeleteUser.ts
@@ -1,0 +1,6 @@
+import { setUserToNotActive } from '../utils/firestore/users';
+import { AuthOnDeleteHandler } from '../types/functions';
+
+export const onDeleteUserHandler: AuthOnDeleteHandler = async ({ uid }) => {
+  await setUserToNotActive(uid);
+};
