+++ packages/functions/src/handlers/onUpdateToken.ts
@@ -1,6 +1,6 @@
 import { TokenData } from '@yukukuru/types';
-import { setUserToActive, setUserToNotActive } from '../../utils/firestore/users';
-import { FirestoreOnUpdateHandler } from '../../types/functions';
+import { setUserToActive, setUserToNotActive } from '../utils/firestore/users';
+import { FirestoreOnUpdateHandler } from '../types/functions';
 
 export const onUpdateTokenHandler: FirestoreOnUpdateHandler = async ({ after }) => {
   const { twitterAccessToken = null, twitterAccessTokenSecret = null, twitterId = null } = after.data() as TokenData;
